requirements.txt
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==1.16.0
exceptiongroup==1.2.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pluggy==1.4.0
psycopg2==2.9.9
pymongo==3.6.0
PySocks==1.7.1
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.25
tomli==2.0.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
webdriver-manager==4.0.1
Werkzeug==3.0.1
wsproto==1.2.0
